tests/wpt/web-platform-tests/app-history/currentchange-event/currentchange-app-history-navigate-same-doc.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = t.step_timeout(resolve, 0));

  let oncurrentchange_called = false;
  appHistory.oncurrentchange = t.step_func(e => {
    oncurrentchange_called = true;
    assert_equals(e.from, appHistory.entries()[0]);
    assert_equals(e.navigationType, "push");
    assert_equals(appHistory.current.index, 1);
  });
  let result = appHistory.navigate("#foo");
  assert_true(oncurrentchange_called);
  await result.committed;
}, "AppHistoryCurrentChangeEvent fires for appHistory.navigate()");
</script>
